t1: IO.foreach` returns the list of results
